worked on unit tests a bit (compute/create output folder, moved writer_test to tests/)
